Scrolling saga: Part 2

+++ util/math.h
@@ -1,7 +1,7 @@
 #ifndef UTIL_MATH_H
 #define UTIL_MATH_H
 
-#define MAX(x, y) (x > y ? x : y)
-#define MIN(x, y) (x < y ? x : y)
+#define MAX(x, y) ((int)(x) > (int)(y) ? x : y)
+#define MIN(x, y) ((int)(x) < (int)(y) ? x : y)
 
 #endif // UTIL_MATH_H
